Add UI to select a CSV file (no backend yet)
New model to give instructions (still needs a “clear all other data
checkbox”).
Let’s you select a file which is cool. :)  Does nothing on the backend
yet.

Also started using data- to show the modals.  This is a change I should
make on the main branch as well.

Added a placeholder button for a team assignments page (a feature that
will come much later).